+++ system/main_cgo_test.go
@@ -1,7 +1,0 @@
-//go:build cgo
-
-package system_test
-
-import (
-	_ "github.com/larsartmann/go-cqrs-lite/metaengine/duckdbengine/v4" // registers "duckdb"
-)